# api-docs/v1/openapi.yaml
openapi: 3.0.3
info:
  title: OMS API
  version: 1.0.0
  description: |
    Versioned Order Management API with dev-friendly docs.
    Use Swagger "Authorize" to test protected routes. Dev auth format:
    `Authorization: Bearer dummy.<role>.<userId>` (roles: sales, admin)
servers:
  - url: /api
    description: API namespace (most endpoints)
  - url: /
    description: Root (also exposes /v1/* and /health)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # dummy today; real token later
  headers:
    X-Request-Id:
      description: Correlates server logs with this response.
      schema:
        type: string
    X-RateLimit-Limit:
      description: Max requests allowed in the current window.
      schema:
        type: integer
        example: 100
    X-RateLimit-Remaining:
      description: Requests left in the current window.
      schema:
        type: integer
        example: 97
    X-RateLimit-Reset:
      description: Unix epoch seconds when the window resets.
      schema:
        type: integer
        format: int64
        example: 1723199999
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: Missing or invalid token
    ValidationError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: 'limit'
                  code:
                    type: string
                    example: 'invalid_type'
                  message:
                    type: string
                    example: 'Expected number, received string'
...
    CreateRuleSet:
      type: object
      required:
        [name, effectiveFrom, unitPriceCents, unitWeightKg, shipRatePerKgKm, shippingMaxRatio]
      properties:
        name:
          type: string
          example: 'Q4-2025'
        effectiveFrom:
          type: string
          format: date-time
          example: '2025-08-10T00:00:00Z'
        unitPriceCents:
          type: integer
          minimum: 0
          example: 15000
        unitWeightKg:
          type: number
          exclusiveMinimum: 0
          example: 0.365
        shipRatePerKgKm:
          type: number
          minimum: 0
          example: 0.01
        shippingMaxRatio:
          type: number
          minimum: 0
          maximum: 1
          example: 0.15
    DiscountTier:
      type: object
      required: [threshold, pct]
      properties:
        threshold:
          type: integer
          minimum: 1
          example: 25
        pct:
          type: integer
          minimum: 0
          maximum: 100
          example: 5
        priority:
          type: integer
          minimum: 0
          example: 0
  responses:
    UnprocessableEntity:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    Unauthorized:
      description: Unauthorized
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                error:
                  code: UNAUTHORIZED
                  message: 'Missing or invalid token'
    Forbidden:
      description: Forbidden
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                error:
                  code: FORBIDDEN
                  message: 'Insufficient role'
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            limited:
              value:
                error:
                  code: RATE_LIMITED
                  message: 'Too many requests'
    NotFound:
      description: Resource not found
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                error:
                  code: NOT_FOUND
                  message: 'Resource not found'
paths:
...
  /v1/admin/pricing/rulesets:
    post:
      tags: [Admin, Pricing]
      summary: Create pricing rule set
      description: |
        Create a rule set with effective date. If effective now, future verifications use it immediately.
      operationId: createPricingRuleSet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleSet'
            examples:
              now:
                value:
                  name: 'Q4-2025'
                  effectiveFrom: '2025-08-10T00:00:00Z'
                  unitPriceCents: 15000
                  unitWeightKg: 0.365
                  shipRatePerKgKm: 0.01
                  shippingMaxRatio: 0.15
      responses:
        '201':
          description: Created
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    example: 'a1b2c3d4-e5f6-7890-abcd-ef1234567890'
                  effectiveFrom:
                    type: string
                    format: date-time
                    example: '2025-08-10T00:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/admin/pricing/rulesets/{id}/tiers:
    put:
      tags: [Admin, Pricing]
      summary: Replace discount tiers for a rule set
      description: |
        Replace all discount tiers for the specified rule set with the provided tiers.
      operationId: replacePricingTiers
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          example: 'a1b2c3d4-e5f6-7890-abcd-ef1234567890'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tiers]
              properties:
                tiers:
                  type: array
                  items:
                    $ref: '#/components/schemas/DiscountTier'
            examples:
              basic:
                value:
                  tiers:
                    - threshold: 25
                      pct: 5
                    - threshold: 50
                      pct: 10
                    - threshold: 100
                      pct: 15
                    - threshold: 250
                      pct: 20
      responses:
        '200':
          description: Replaced
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                type: object
                properties:
                  ruleSetId:
                    type: string
                    format: uuid
                    example: 'a1b2c3d4-e5f6-7890-abcd-ef1234567890'
                  tiers:
                    type: integer
                    example: 4
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'